move files, simplify code
