+++ pipes-and-filters/src/main/scala/asso/pipes/pull/Filters.scala
